Redirect file that Daniel Beck noted has been replaced
https://github.com/jenkins-infra/jenkins.io/pull/4612 includes a
detailed review by Daniel where he recommended that Meg McRoberts
proceed with the removal of the file that references this location.

The content on the page is well covered by the destination of the
redirect and the destination of the redirect is being updated and
corrected as needed for further improvements.