Delete spam from JA archive
Removed the "Content removed" section from the JA archive since it was
entirely spam content.

I searched the rest of the archive with `git grep -i -E -w` for many
of the words mentioned in the deleted spam files and confirmed that the
spam was limited to the JA subdirectory.

I checked the images in the deleted files and found only the 1 image
that needed to be deleted.

I checked the images in content/JA/attachments with my local file explorer
and found no other images that needed to be deleted.